Merge to Master

+++ src/app/admin.service.ts
@@ -14,35 +14,39 @@ export class AdminService {
       .toPromise();
   }
 
-  getTables() {
-    return this.httpClient.get(environment.API_URL + 'tables');
+  getRestaurantByUserId(userId: number) {
+    return this.httpClient.get(`${environment.API_URL}/admin/getRestaurantByUserId?userId=${userId}`);
+  }
+
+  getTables(restaurantId) {
+    return this.httpClient.get(environment.API_URL + '/admin/getTablesByResId?restaurantId=' + restaurantId);
   }
   editTable(data) {
-    return this.httpClient.post(environment.API_URL + 'tables', data);
+    return this.httpClient.put(environment.API_URL + '/admin/tables', data);
   }
   deleteTable(tableId) {
-    return this.httpClient.delete(environment.API_URL + 'tables/' + tableId);
+    return this.httpClient.delete(environment.API_URL + `/admin/tables/${tableId}` );
   }
   createTable(data) {
-    return this.httpClient.post(environment.API_URL + 'tables', data);
+    return this.httpClient.post(environment.API_URL + '/admin/tables', data);
   }
 
   getMenu(id) {
     return this.httpClient.get(
-      environment.API_URL + 'getMenu?restaurantId=' + id
+      environment.API_URL + '/admin/getMyMenu?restaurantId=' + id
     );
   }
   createMenu(data) {
-    return this.httpClient.post(environment.API_URL + 'menu', data);
+    return this.httpClient.post(environment.API_URL + '/admin/menu', data);
   }
   deleteMenu(menuId) {
-    return this.httpClient.delete(environment.API_URL + 'menu/' + menuId);
+    return this.httpClient.delete(environment.API_URL + '/admin/menu/' + menuId);
   }
   updatemenu(data) {
-    return this.httpClient.post(environment.API_URL + 'menu', data);
+    return this.httpClient.put(environment.API_URL + '/admin/menu', data);
   }
 
   getAllOrders(id) {
-    return this.httpClient.get(environment.API_URL + 'orferByResId/' + id);
+    return this.httpClient.get(environment.API_URL + '/admin/ordersByRestaurantId?restaurantId=' + id);
   }
 }
